+++ server/src/main/java/com/blocklang/core/service/impl/PropertyServieImpl.java
@@ -3,6 +3,7 @@ package com.blocklang.core.service.impl;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
 import com.blocklang.core.dao.PropertyDao;
@@ -18,6 +19,7 @@ public class PropertyServieImpl implements PropertyService {
 	public PropertyDao propertyDao;
 	
 	@Override
+	@Cacheable(value = "cm_properties")
 	public Optional<String> findStringValue(String key) {
 		Optional<CmProperty> property = propertyDao.findByKeyAndParentIdAndValid(key, ROOT_ID, true);
 		return property.map(CmProperty::getValue);